requirements.txt
tensorflow
streamlit
numpy
pillow
matplotlib
seaborn
scikit-learn
XlsxWriter
pandas